Ensure trampoline destination stores precede their execution
---

I haven't actually seen a bug from this (or even compiled the code, for
that matter), I just found this when trying to sort out how the aarch64
port handles the modification of instruction streams that can be
executing concurrently.  I'm actually really just interested in how we
handle this in RISC-V land, but as I think I've found a bug I figured I
might as well send it along -- at a bare minimum I'll be able to figure
out how my understanding of the arm64 memory model is broken.

From my understanding of this code, it's possible for a code sequence
that looks something like (forgive my pseudo-arm64 assembly)

    j TARGET
    ...
    ret
    ...
    trampoline:
    ldr Rtmp, target
    jr Rtmp
    target:
    .long 0

to end up being modified to look something like

    j trampoline
    ...
    ret
    ...
    trampoline:
    ldr Rtmp, target
    jr Rtmp
    target:
    .long NEW_TARGET

This modification happens in two phases: first the target address is
written, and then the jump is pointed at the trampoline.  These are
currently ordered via a "dmb ish", and while that may be sufficient for
other trampoline update requirements it doesn't enforce any ordering WRT
instruction fetches.  That means that the jump to the trampoline may be
fetched before the new target address is guaranteed to be visible to the
remote core.